+++ src/main/java/com/example/sakilaapi/service/film/FilmService.java
@@ -1,23 +1,21 @@
 package com.example.sakilaapi.service.film;
 
 import com.example.sakilaapi.dto.FilmDto;
-import com.example.sakilaapi.model.Film;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface FilmService {
     List<FilmDto> getAllFilms();
-    FilmDto getFilmById(Short id);
+    FilmDto getFilmById(Integer id);
     List<FilmDto> searchFilms(Integer releaseYear, List<Integer> categoryIds);
-    List<FilmDto> getFilmsByActor(Short actorId);
+    List<FilmDto> getFilmsByActor(Integer actorId);
     List<FilmDto> getAvailableFilms();
     FilmDto saveFilm(FilmDto filmDto);
 
-    FilmDto updateFilm(Short id,FilmDto filmDto);
+    FilmDto updateFilm(Integer id,FilmDto filmDto);
 
     void deleteFilm(FilmDto filmDto);
 
-    void deleteFilmById(Short id);
+    void deleteFilmById(Integer id);
 }
