+++ src/main/java/com/example/sakilaapi/service/film/FilmServiceImpl.java
@@ -2,7 +2,6 @@ package com.example.sakilaapi.service.film;
 
 import com.example.sakilaapi.dto.FilmDto;
 import com.example.sakilaapi.mapper.FilmMapper;
-import com.example.sakilaapi.model.Actor;
 import com.example.sakilaapi.model.Film;
 import com.example.sakilaapi.repository.FilmRepository;
 import jakarta.ws.rs.NotFoundException;
@@ -24,7 +23,7 @@ public class FilmServiceImpl implements FilmService{
     }
 
     @Override
-    public FilmDto getFilmById(Short id) {
+    public FilmDto getFilmById(Integer id) {
         Optional<Film> optionalFilm = filmRepository.getById(id);
         if (!optionalFilm.isPresent()){
             return filmMapper.toDto(optionalFilm.get());
@@ -38,7 +37,7 @@ public class FilmServiceImpl implements FilmService{
     }
 
     @Override
-    public List<FilmDto> getFilmsByActor(Short actorId) {
+    public List<FilmDto> getFilmsByActor(Integer actorId) {
         return filmMapper.toDto(filmRepository.findFilmsByActorId(actorId));
     }
 
@@ -53,7 +52,7 @@ public class FilmServiceImpl implements FilmService{
     }
 
     @Override
-    public FilmDto updateFilm(Short id,FilmDto filmDto) {
+    public FilmDto updateFilm(Integer id,FilmDto filmDto) {
         Optional<Film> optionalFilm = filmRepository.getById(id);
         if (!optionalFilm.isPresent())
             throw new NotFoundException("Can't get the film with this id");
@@ -70,7 +69,7 @@ public class FilmServiceImpl implements FilmService{
     }
 
     @Override
-    public void deleteFilmById(Short id) {
+    public void deleteFilmById(Integer id) {
         filmRepository.deleteById(id);
     }
 }
